refactor: RowMapper bind 패키지로 이동
