+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -2,7 +2,7 @@ package com.techcourse.dao;
 
 import com.interface21.jdbc.core.JdbcTemplate;
 import com.interface21.jdbc.core.NamedJdbcTemplate;
-import com.interface21.jdbc.core.RowMapper;
+import com.interface21.jdbc.bind.RowMapper;
 import com.techcourse.domain.User;
 import java.util.List;
 import java.util.Optional;
